+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Linux: New DRM/KMS screen grabber with plane-based capture - not feature complete yet
 - Logging/Tracing: Introduced qlogging categories to enable dynamic tracing
 - Home Assistant: Dynamically set brightness for higher dynamic range (#1922)
+- Add DRM_FORMAT_RGB565 format to DRM frame grabber
 
 ---